Rename manifestation.py to manifestations.py for consistency with route names
