+++ web/views/manifestations.py
